requirements.txt
pyspark==3.5.4
delta-spark==3.3.0